requirements.txt
# API y Web
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-dotenv==1.0.0
pyjwt==2.6.0
jinja2==3.1.2

# Base de datos
psycopg2-binary==2.9.5
redis==4.5.4
alembic==1.10.2

# Computer Vision
opencv-python==4.7.0.72
ultralytics==8.0.6
torch>=1.7.0
torchvision>=0.8.1
numpy==1.24.2
pillow==9.4.0

# Procesamiento
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.7.1
pandas==1.5.3

# Utilidades
pyyaml==6.0
python-multipart==0.0.6
tenacity==8.2.2
aiofiles==23.1.0
